+++ src/dqn/pytorch_dqn_atari.py
@@ -62,7 +62,7 @@ def make_env(env_id, capture_video=False):
 class ReplayBuffer:
     def __init__(self, buffer_size, batch_size, state_dim, device):
         self.state_buffer = np.zeros((buffer_size,) + state_dim, dtype=np.int8)
-        self.action_buffer = np.zeros((buffer_size), dtype=np.int64)
+        self.action_buffer = np.zeros((buffer_size), dtype=np.int32)
         self.reward_buffer = np.zeros((buffer_size), dtype=np.float32)
         self.flag_buffer = np.zeros((buffer_size), dtype=np.float32)
